Update parser instructions/readme file
