+++ src/util/untokenize.py
@@ -1,4 +1,17 @@
-sentence = "Ayn Ghazal -LRB- Ain Ghazal , ʿayn ġazāl عين غزال -RRB- is a neolithic archaeological site located in metropolitan Amman , Jordan , about 2 km north-west of Amman Civil Airport ."
+# Copyright 2018 Amazon Research Cambridge
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
 
 replacements = {
     "-LRB-":"(",
