+++ src/util/wiki.py
@@ -13,13 +13,13 @@
 # limitations under the License.
 
 
-import os
 import logging
-from dataset.jobs.test3 import untokenize
+import os
 
 import botocore
 
 from persistence.s3_persistence import S3Writer
+from util.untokenize import untokenize
 
 logger = logging.getLogger(__name__)
